pyproject.toml
[project]
name = "alphaagents"
version = "0.1.0"
description = "Multi-agent AI system for tax-loss harvesting optimization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core dependencies
    "crewai>=0.177.0",
    "fastapi>=0.104.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.33",
    "pypdf>=6.0.0",
    "python-multipart>=0.0.20",
    "streamlit>=1.49.1",
    # Financial data
    "yfinance>=0.2.65",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    # Vector store
    "chromadb>=0.4.0",
    # Environment management
    "python-dotenv>=1.0.0",
    # Async support
    "uvicorn[standard]>=0.24.0",
    # API validation
    "pydantic>=2.0.0",
    # Monitoring & metrics
    "prometheus-client>=0.18.0",
    # Web scraping
    "tavily-python>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pylint>=2.17.0",
    "radon>=6.1.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
]

monitoring = [
    "prometheus-client>=0.18.0",
]

docker = [
    "docker>=6.0.0",
    "docker-compose>=1.29.0",
]
